Move columnName function to ui.js
